sysbuild: update image specific names to use DEFAULT_IMAGE
The image name is hard coded for dependency configuration, but when
using sysbuild the image name of the default image is based on the
name of the folder. This means that if sample is copied elsewhere for
local modification the hard coded name may no longer match the default
image name.

Improve this by using DEFAULT_IMAGE variable instead of hard coded name.

Jira: NCSDK-28028

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>